sum of a list

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -5,7 +5,7 @@ Module that returns the sum of a list
 from typing import List, Union
 
 
-def sum_mixed_list(mxd_list: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_list: List[Union[int, float]]) -> float:
     """Function that returns the sum of floats in a list"""
 
     return sum(mxd_list)
